electron-app/src/python/requirements.txt
sounddevice>=0.4.6
numpy>=1.26.4
deepgram-sdk>=2.12.0
asyncio-mqtt>=0.16.1
